# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sample crud application using vuetift 3, go lang, postgres",
        "title": "crud application swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ishmam abir",
            "url": "https://github.com/ishmamabir",
            "email": "ishmam.cse@gmail.com"
        },
        "license": {
            "name": "apache 2.0",
            "url": "http://www.apache.org/licenses/license-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9080",
    "basePath": "/",
    "paths": {
        "/users": {
            "get": {
                "description": "get User List",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "post": {
                "description": "Takes CreateUser Request",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "create User",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "get User By Id",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "put": {
                "description": "Update User Request",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "delete": {
                "description": "Takes CreateUser Request",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                }
            }
        }
    }
}